fixed issue causing chart history to erase
updated the add and remove route/functions to load the data.json file which contains the chart's historical data, and appends the new data when removing or adding an item. This fixed the issue where the add and remove functions were overwriting the data.json. This is a mitigation until this app is integrated into an actual database.